Add link to leaderboard, improve header responsiveness
