--- app/src/components/shared/Account/style.module.scss
@@ -1,26 +1,0 @@
-.account {
-	position: relative;
-
-	&__user {
-		padding-inline: 0 !important;
-	}
-
-	&__dropdown {
-		right: 0;
-		display: flex;
-		flex-direction: column;
-		gap: 12px;
-		min-width: 200px;
-
-		.links {
-			display: flex;
-			flex-direction: column;
-			gap: 4px;
-
-			&__link {
-				justify-content: flex-start;
-				padding-inline: 0;
-			}
-		}
-	}
-}
