+++ app/src/components/shared/HeaderDropdown/style.module.scss
@@ -1,0 +1,74 @@
+@use '@/scss/breakpoints' as *;
+
+.dropdown {
+	position: relative;
+
+	&__toggle {
+		@media screen and (width > $screen-md) {
+			padding-inline: 0 !important;
+		}
+
+		.toggle-icon {
+			width: 20px;
+			height: 20px;
+		}
+	}
+
+	&__dropdown {
+		right: 0;
+		display: flex;
+		flex-direction: column;
+		gap: 12px;
+		min-width: 200px;
+
+		.links {
+			display: flex;
+			flex-direction: column;
+			gap: 4px;
+			margin: 0;
+			padding: 0;
+
+			&__link {
+				display: block;
+
+				.button {
+					justify-content: flex-start;
+					padding-inline: 0;
+					width: 100%;
+
+					@at-root {
+						[data-theme='light'] .button[aria-current='page'] {
+							color: var(--primary-dark);
+						}
+
+						[data-theme='dark'] .button[aria-current='page'] {
+							color: var(--primary);
+						}
+					}
+				}
+			}
+		}
+
+		.user-info {
+			display: flex;
+			align-items: center;
+			gap: 8px;
+		}
+	}
+}
+
+.desktop {
+	display: none;
+
+	@media screen and (width > $screen-md) {
+		display: contents;
+	}
+}
+
+.mobile {
+	display: contents;
+
+	@media screen and (width > $screen-md) {
+		display: none;
+	}
+}
